Switched to webpack and typescript
Switched the project to use webpack
and typescript, so now the dist folder
holds the javascript that gets loaded
by the manifest
