+++ manifest.json
@@ -8,7 +8,7 @@
                 "<all_urls>"
             ],
             "js": [
-                "./src/script.js"
+                "./dist/script.js"
             ]
         }
     ],
