+++ options.html
@@ -6,7 +6,7 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <title>Document</title>
         <link rel="stylesheet" href="./src/css/style.css" />
-        <script src="./src/options.js" defer></script>
+        <script src="./dist/options.js" defer></script>
     </head>
     <body>
         <div class="container">
@@ -20,3 +20,4 @@
         </div>
     </body>
 </html>
+
