+++ package.json
@@ -1,10 +1,14 @@
 {
-  "name": "mediacontrols",
+  "name": "my-webpack-project",
   "version": "1.0.0",
-  "description": "",
+  "description": "My webpack project",
   "main": "service_worker.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "webpack --mode=production --node-env=production",
+    "build:dev": "webpack --mode=development",
+    "build:prod": "webpack --mode=production --node-env=production",
+    "watch": "webpack --watch"
   },
   "repository": {
     "type": "git",
@@ -18,6 +22,14 @@
   },
   "homepage": "https://github.com/Sandro-Enzo/mediacontrols#readme",
   "devDependencies": {
-    "@types/chrome": "0.0.193"
+    "@types/chrome": "0.0.193",
+    "@webpack-cli/generators": "^2.5.0",
+    "prettier": "^2.7.1",
+    "sass": "^1.49.8",
+    "sass-loader": "^12.6.0",
+    "ts-loader": "^9.2.6",
+    "typescript": "^4.6.2",
+    "webpack": "^5.73.0",
+    "webpack-cli": "^4.10.0"
   }
 }
