+++ src/normal.ts
@@ -1,0 +1,4 @@
+export default class Normal extends Option {
+    visibilityChange() {}
+}
+
