+++ tsconfig.json
@@ -1,0 +1,11 @@
+{
+    "compilerOptions": {
+        "allowSyntheticDefaultImports": true,
+        "noImplicitAny": true,
+        "module": "es6",
+        "target": "es5",
+        "allowJs": true
+    },
+    "files": ["src/script.ts", "src/options.ts"]
+}
+
